+++ src/refactoring/App.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { CartPage } from "./components/cart/CartPage.tsx";
-import { AdminPage } from "./components/admin/AdminPage.tsx";
+import { AdminPage } from "./pages/AdminPage.tsx";
 import { useCoupons, useProducts } from "./hooks";
 import { initialProducts, initialCoupons } from "./shared/constants";
